+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## [Unreleased]
 
+## [0.0.5] - 2018-11-06
+### Changed
+- Store enhancer support
+
+## [0.0.4] - 2018-11-05
+### Changed
+- Redux-like createStore
+- Update README
+
+## [0.0.3] - 2018-11-04
+### Added
+- Reducer support
+
 ## [0.0.2] - 2018-10-30
 ### Changed
 - Update README